requirements.txt
azure-functions==1.24.0
certifi==2025.10.5
charset-normalizer==3.4.4
exceptiongroup==1.3.0
idna==3.11
iniconfig==2.1.0
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
requests==2.32.5
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
